group by element in edit page

+++ src/main/react-frontend/src/components/Orders.jsx
@@ -38,6 +38,8 @@ function Orders() {
 
 
     const handleDelete = (orderId) => {
+        if(!window.confirm("êtes-vous sûr ?")) return;
+
         OrderService.removeOrder(orderId)
             .then(
                 () => {
